+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 - Require Laravel 8.
 
+## Added
+- AMQP config file
+
 ## [0.2.13] - 2020-08-03
 
 ### Added