+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+- Include [butler-audit](https://github.com/glesys/butler-audit)
 
 ## [0.3.0] - 2020-10-08