+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.3.1] - 2020-11-09
+
 ## Added
 - Include [butler-audit](https://github.com/glesys/butler-audit)